# ticketing/documentation/API_documentation.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "Ticketing API",
    "version": "1.0.0",
    "description": "API endpoints for managing the ticketing service"
  },
  "paths": {
...
    "/API/customers/tickets/{idTicket}/messages": {
      "get": {
        "summary": "Get tickets with messages by customer ID",
        "operationId": "getTicketsWithMessagesByCustomerId",
        "parameters": [
          {
            "name": "idTicket",
            "in": "path",
            "description": "Ticket ID",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful operation",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/MessageDTO"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ]
      },
      "post": {
        "summary": "Add a new message to a ticket",
        "operationId": "addMessage",
        "parameters": [
          {
            "name": "idTicket",
            "in": "path",
            "description": "Ticket ID",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "requestBody": {
          "description": "Message data",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MessageDTO"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Message added successfully"
          },
          "401": {
            "description": "Unauthorized"
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ]
      }
    },
...
    "/API/manager/tickets": {
      "get": {
        "summary": "Get tickets by status",
        "operationId": "getTicketsByStatus",
        "parameters": [
          {
            "name": "status",
            "in": "query",
            "description": "Ticket status",
            "required": false,
            "schema": {
              "$ref": "#/components/schemas/TicketStatus"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful operation",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/TicketDTO"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ]
      }
    },
    "/API/manager/customers": {
      "get": {
        "summary": "Get a customer by ID",
        "operationId": "getCustomer",
        "parameters": [
          {
            "name": "id",
            "in": "query",
            "description": "Customer ID",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful operation",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CustomerDTO"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ]
      }
    },
    "/API/manager/experts": {
      "get": {
        "summary": "Get an expert by ID",
        "operationId": "getExpert",
        "parameters": [
          {
            "name": "id",
            "in": "query",
            "description": "Expert ID",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful operation",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EmployeeDTO"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ]
      }
    },
    "/API/manager/tickets/{idTicket}/messages": {
      "get": {
        "summary": "Get messages for a ticket",
        "operationId": "getTicketsWithMessagesByCustomerId",
        "parameters": [
          {
            "name": "idTicket",
            "in": "path",
            "description": "Ticket ID",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful operation",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/MessageDTO"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ]
      }
    },
    "/API/manager/tickets/{idTicket}/assign": {
      "put": {
        "summary": "Assign a ticket to an expert",
        "operationId": "ticketAssign",
        "parameters": [
          {
            "name": "idTicket",
            "in": "path",
            "description": "Ticket ID",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          },
          {
            "name": "expert",
            "in": "query",
            "description": "Expert ID",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "priority",
            "in": "query",
            "description": "Priority level",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/PriorityLevel"
            }
          }
        ],
        "responses": {
          "202": {
            "description": "Ticket assigned successfully"
          },
          "401": {
            "description": "Unauthorized"
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ]
      }
    },
    "/API/messages/{messageId}/attachments": {
      "get": {
        "summary": "Get attachments for a message",
        "operationId": "getMessageAttachments",
        "parameters": [
          {
            "name": "messageId",
            "in": "path",
            "description": "Message ID",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful operation",
            "content": {
              "image/jpeg": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/AttachmentDTO"
                  }
                }
              },
              "image/png": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/AttachmentDTO"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ]
      },
      "post": {
        "summary": "Add attachments to a message",
        "operationId": "addAttachments",
        "parameters": [
          {
            "name": "messageId",
            "in": "path",
            "description": "Message ID",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "properties": {
                  "attachment": {
                    "type": "array",
                    "items": {
                      "type": "string",
                      "format": "binary"
                    }
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Attachments added successfully"
          },
          "401": {
            "description": "Unauthorized"
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ]
      }
    },
    "/API/attachments": {
      "get": {
        "summary": "Get an attachment",
        "operationId": "getAttachment",
        "parameters": [
          {
            "name": "id",
            "in": "query",
            "description": "Attachment ID",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful operation",
            "content": {
              "image/jpeg": {
                "schema": {
                  "type": "string",
                  "format": "binary"
                }
              },
              "image/png": {
                "schema": {
                  "type": "string",
                  "format": "binary"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ]
      }
    },
    "/API/messages": {
      "get": {
        "summary": "Get messages by ticket ID",
        "operationId": "getMessagesByIdTickets",
        "parameters": [
          {
            "name": "ticket",
            "in": "query",
            "description": "Ticket ID",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful operation",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/MessageDTO"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ]
      }
    },
    "/API/messages/{messageId}": {
      "put": {
        "summary": "Edit a message",
        "operationId": "editMessage",
        "parameters": [
          {
            "name": "messageId",
            "in": "path",
            "description": "Message ID",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "string"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Message edited successfully"
          },
          "401": {
            "description": "Unauthorized"
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ]
      }
    },
...
    "/API/tickets/{idCustomer}/messages": {
      "get": {
        "summary": "Get tickets with messages by customer ID",
        "operationId": "getTicketsWithMessagesByCustomerId",
        "parameters": [
          {
            "name": "idCustomer",
            "in": "path",
            "description": "Customer ID",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful operation",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/TicketWithMessagesDTO"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ]
      }
    },
    "/API/tickets/{ticketId}/close": {
      "put": {
        "summary": "Close a ticket",
        "operationId": "ticketClosing",
        "parameters": [
          {
            "name": "ticketId",
            "in": "path",
            "description": "Ticket ID",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Ticket closed successfully"
          },
          "401": {
            "description": "Unauthorized"
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ]
      }
    },
    "/API/tickets/{ticketId}/messages": {
      "get": {
        "summary": "Get messages of a ticket",
        "operationId": "getMessages",
        "parameters": [
          {
            "name": "ticketId",
            "in": "path",
            "description": "Ticket ID",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful operation",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/MessageDTO"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ]
      }
    },
    "/API/tickets/{ticketId}/status": {
      "get": {
        "summary": "Get status of a ticket",
        "operationId": "getStatus",
        "parameters": [
          {
            "name": "ticketId",
            "in": "path",
            "description": "Ticket ID",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful operation",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TicketStatus"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ]
      }
    }
  },
  "components": {
    "schemas": {
      "CustomerDTO": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "first_name": {
            "type": "string"
          },
          "last_name": {
            "type": "string"
          },
          "email": {
            "type": "string",
            "format": "email"
          },
          "dob": {
            "type": "string",
            "format": "date"
          },
          "address": {
            "type": "string"
          },
          "phone_number": {
            "type": "string"
          }
        }
      },
      "MessageDTO": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "body": {
            "type": "string"
          },
          "date": {
            "type": "string",
            "format": "date-time"
          },
          "listOfAttachments": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/AttachmentDTO"
            }
          },
          "expert": {
            "type": "string",
            "format": "uuid"
          }
        }
      },
      "TicketDTO": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "dateTime": {
            "type": "string",
            "format": "date-time"
          },
          "title": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "priority": {
            "$ref": "#/components/schemas/PriorityLevel"
          },
          "product": {
            "$ref": "#/components/schemas/ProductDTO"
          }
        }
      },
      "TicketWithMessagesDTO": {
        "type": "object",
        "properties": {
          "ticket": {
            "$ref": "#/components/schemas/TicketDTO"
          },
          "listOfMessage": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/MessageDTO"
            }
          }
        }
      },
      "EmployeeDTO": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "description": "The ID of the employee (optional)"
          },
          "first_name": {
            "type": "string",
            "description": "The first name of the employee"
          },
          "last_name": {
            "type": "string",
            "description": "The last name of the employee"
          },
          "email": {
            "type": "string",
            "description": "The email address of the employee"
          },
          "type": {
            "type": "string",
            "enum": [
              "ROLE_Manager",
              "ROLE_Expert",
              "ROLE_Client"
            ],
            "description": "The role or type of the employee"
          }
        }
      },
      "AttachmentDTO": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64",
            "description": "The ID of the attachment (optional)"
          },
          "attachment": {
            "type": "string",
            "format": "binary",
            "description": "The binary data of the attachment"
          }
        }
      },
      "ProductDTO": {
        "type": "object",
        "properties": {
          "ean": {
            "type": "string",
            "description": "The EAN of the product"
          },
          "name": {
            "type": "string",
            "description": "The name of the product"
          },
          "brand": {
            "type": "string",
            "description": "The brand of the product"
          }
        }
      },
      "TicketStatus": {
        "type": "string",
        "enum": [
          "OPEN",
          "CLOSED",
          "RESOLVED",
          "IN_PROGRESS",
          "REOPENED"
        ],
        "description": "The status of a ticket"
      },
      "PriorityLevel": {
        "type": "string",
        "enum": [
          "LOW",
          "MEDIUM",
          "HIGH",
          "NOT_ASSIGNED"
        ],
        "description": "The priority level of a ticket"
      }
    },
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  }
}